requirements.txt
asn1crypto==0.24.0
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
Authlib==0.9
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3.1
flake8==3.5.0
Flask==1.0.2
Flask-JSONRPC==0.3.1
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.3.0
pathlib2==2.3.2
pluggy==0.7.1
psycopg2-binary==2.7.6.1
py==1.6.0
pycodestyle==2.3.1
pycparser==2.19
pyflakes==1.6.0
pylint==2.1.1
pytest==3.8.1
pytest-flask==0.12.0
requests==2.20.0
six==1.11.0
typed-ast==1.1.0
urllib3==1.23
Werkzeug==0.14.1
wrapt==1.10.11